requirements.txt
streamlit==1.33.0
yfinance==0.2.37
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.1
scikit-learn==1.2.2
tensorflow-cpu==2.11.0
